% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{addMotif}
\alias{addMotif}
\alias{addMotif.default}
\alias{addMotif.mT1}
\title{Add Motif}
\usage{
addMotif(object, motif, ...)

\method{addMotif}{default}(object, motif, ...)

\method{addMotif}{mT1}(object, motif, ...)
}
\arguments{
\item{object}{A object with an addMotif method}

\item{motif}{A Atomic Character containing IUPAC characters}

\item{...}{Variables to pass to object.}
}
\value{
mT1 object with new motif
}
\description{
Takes a motif and creates a new object based on the class of object
}
\examples{
new<-addMotif(mT1_sampleMT1,"CACCTG")
}
