% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{btest}
\alias{btest}
\title{binomial test}
\usage{
btest(k, n, p, min = 10)
}
\arguments{
\item{k}{count}

\item{n}{total}

\item{p}{prob}

\item{min}{minimum k cut off}
}
\value{
p-value
}
\description{
A wrapper around binom.test that ensures that n!=0 and that k is at
least a minimum number.
}

