% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{combHeights}
\alias{combHeights}
\title{Combination Heights}
\usage{
combHeights(x, ...)
}
\arguments{
\item{x}{seqence of possible values represented as indicies}

\item{...}{vectors of values to be mapped}
}
\description{
Determines the number of occurances in each member ... along x.
}
\examples{
x<-seq(20)
y<-runif(200,1,20)
par(mfrow=c(1,2))
plot(y)
plot(x,combHeights(x,y)[[1]])
}

