% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{complement}
\alias{complement}
\title{Complement}
\usage{
complement(string)
}
\arguments{
\item{string}{the input string of nucleotides}
}
\value{
The composite string
}
\description{
Takes a string of neucleotides (ACTG) and returns the complement
(TGAC). This works on strings only, for DNAStringSet please refer
to the `Biostrings::complement` function. To transform IUPAC characters
into neucleotide form refer to IUPACtoBase.
}
\examples{
mT1:::complement(mT1:::IUPACtoBase("CANNTG"))
mT1:::complement("CA[AC]TT[ACGT]GG")
}

