% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{diffMotif}
\alias{diffMotif}
\title{Motif Difference}
\usage{
diffMotif(fasta, motifs, mloc, cloc, i = 1, j = 2, min = 200,
  combine = "Merged")
}
\arguments{
\item{fasta}{DNAStringSet}

\item{motifs}{list of motif Character Atomics}

\item{mloc}{list of locations of each motif in fasta}

\item{cloc}{list of locations of each compositemotif in fasta}

\item{i}{index 1}

\item{min}{numerical atomic, minimum length of mloc[[i or j]]}

\item{combine}{Merged|First}

\item{2}{index 2}
}
\value{
two columns , the first being the index the second being the
distance
}
\description{
Find the distances between two motifs within a subset of the genome
represented by fasta.
}

