% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{ePD}
\alias{ePD}
\title{ePD}
\usage{
ePD(t1, a, b, width)
}
\arguments{
\item{t1}{width 2 numerical vector of distances}

\item{a}{numerical vector PWM of motif a}

\item{b}{numerical vector PWM of motif b}

\item{width}{width of fasta indicies}
}
\value{
list(hs,mp,p-values)
}
\description{
Determine the pvalue from a vetor of distances and PWMs from the
two motifs being compared
}

