% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fasta}
\alias{fasta}
\title{Fasta}
\format{A DANStringSet with 5670 members:}
\usage{
fasta
}
\description{
The nucleotide values for the genomic locations defined by peaks
}
\keyword{datasets}

