% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{findLocs}
\alias{findLocs}
\title{Find Motif Locations}
\usage{
findLocs(fasta, mloc, cloc, n1, combine = "Merged")
}
\arguments{
\item{fasta}{DNAStringSet}

\item{mloc}{fasta indicies that have  motifs}

\item{cloc}{fasta indicies that have compliment motifs}

\item{n1}{motif name, Atomic Character of only IUPAC characters}

\item{combine}{If two motifs are found on the same index Merged|First}
}
\value{
two columns , the first being the instance the second being the
distance
}
\description{
Finds the location of motifs on each indici of fasta. The results can
be used to build the emperical PDF of motif locations.
}

