% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{getJASPAR}
\alias{getJASPAR}
\title{Get Jaspar}
\usage{
getJASPAR(name)
}
\arguments{
\item{name}{composite motifs IUPAC characters only}
}
\value{
the name of the motif if it came from the jaspar database
}
\description{
To be used in conjuction with the jaspar.RData found in `extdata`.
Pass `getJASPAR` a composite motif from the jaspar database
and it will return the name of the motif.
}
\examples{
# load the required Jaspar motif data
load(system.file("exdata","jaspar.RData",package="mT1"))
getJASPAR(jaspar$names[1])
}

