% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{isIUPAC}
\alias{isIUPAC}
\title{Is IUPAC}
\usage{
isIUPAC(string)
}
\arguments{
\item{string}{An atomic character}
}
\value{
A logical, True if all IUPAC false otherwise
}
\description{
Check if a string contains only iupac characters
}
\examples{
isIUPAC("CANNTG")
isIUPAC("CAXCTG")
}
