% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{IUPACDNA}
\alias{IUPACDNA}
\title{IUPACDNA}
\format{List with 15 members one for each IUPAC charcter :}
\usage{
IUPACDNA
}
\description{
A list with IUPAC names "A C G T R Y S W K M B D H V N" and
base pair values. eg IUPACDNA["A"] -> "A" &
IUPACDNA["N"]->c("A","C","T","G")
}
\keyword{datasets}
