% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{IUPACtoBase}
\alias{IUPACtoBase}
\title{IUPAC to base}
\usage{
IUPACtoBase(char, rl = FALSE)
}
\arguments{
\item{char}{the input string of IUPAC characters}

\item{rl}{flag to return all variants}
}
\value{
Character containing neucleotides + brackets
}
\description{
Transforms an IUPAC dna sequence into a string of neucleotides.
Combo characters are replaced by their compoenents in bracket.
For example CNC -> C[ACGT]C. With the rl flag set to TRUE the
expanded set of sequences will be returnd. For example CNC
would become c("CAC","CCC","CGC","CTC").
}
\examples{
mT1:::IUPACtoBase("HGATAA")
mT1:::IUPACtoBase("CANNTG")
mT1:::IUPACtoBase("CANNTG",TRUE)#!/usr/bin/env R
}

