% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{jaspar}
\alias{jaspar}
\title{Jaspar}
\format{\preformatted{List of 4
 $ jnames: chr [1:519] "YTGTGGTT" "GCCYSAGGC" "CACGTG" "GCGTG" ...
 $ jfid  : chr [1:519] "RUNX1" "TFAP2A" "Arnt" "Ahr::Arnt" ...
 $ jl    : Named int [1:519] 571 105 217 940 0 0 0 35 20 26 ...
  ..- attr(*, "names")= chr [1:519] "YTGTGGTT" "GCCYSAGGC" "CACGTG" "GCGTG" ...
 $ jsublM: chr [1:519] "MCWTNT" "HGATAA" "GGGGA" "TAAWNG" ...
}}
\usage{
jaspar
}
\description{
Jaspar
}
\keyword{datasets}

