% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{main.mT1}
\alias{main.mT1}
\title{main mT1}
\usage{
\method{main}{mT1}(obj)
}
\arguments{
\item{obj}{mT1 object}
}
\value{
data.frame(<motif1>,<motif2>,<max pvalue><loc max pvalue>)
}
\description{
returns the pertanent infromation concerning obj
}
\examples{
load(system.file("extdata","objMT1.RData",package="mT1"))
main(objMT1)
}

