% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{main}
\alias{main}
\title{main}
\usage{
main(x, ...)
}
\description{
returns the pertanent information in an object
}
\examples{
load(system.file("exdata","objMT1.RData",package="mT1"))
main(1)
main(objMT1)
}

