% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{makeTitle.mT1}
\alias{makeTitle.mT1}
\title{Make Title mT1}
\usage{
\method{makeTitle}{mT1}(obj, i)
}
\arguments{
\item{obj}{mT1 object}

\item{i}{indici of obj$combs}
}
\value{
atomic Character with motif info
}
\description{
Generates a title based on the composite motifs of a combination
}

