% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{makeTitle}
\alias{makeTitle}
\alias{makeTitle.mT1}
\title{Make Title}
\usage{
makeTitle(x, ...)

\method{makeTitle}{mT1}(x, i, ...)
}
\arguments{
\item{x}{obj to make title from}

\item{...}{possible properties}

\item{i}{print index}
}
\value{
Atomic Character with motif info
}
\description{
Generates a title to be used for plotting object x
}

