% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{motifDistance}
\alias{motifDistance}
\title{Motif Distance}
\usage{
motifDistance(fasta, motif1, motif2)
}
\arguments{
\item{fasta}{a DNAStringSet}

\item{motif1}{Character of IUPAC characters}

\item{motif2}{Character of IUPAC characters}
}
\value{
two columns , the first being the index the second being the
distance
}
\description{
Find the distances between two motifs within a subset of the genome
represented by fasta.
}
\examples{
## load fasta file
load(system.file("extdata","fasta.RData",package="mT1"))
## Find distances
distances<-motifDistance(fasta,"CANNTG","HGATAA")

## Determine the range of the motif locations and plot the
## frequency
width<-nchar(fasta[1]) # all fasta strings should be the same width
r<-seq(-width+1,width-1)
y<-combHeights(r,distances[,2])[[1]]
plot(r,y,main="CANNTG-HGATAA")
}

