% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{motifPDF}
\alias{motifPDF}
\title{Motif PDF}
\usage{
motifPDF(fasta, motif)
}
\arguments{
\item{fasta}{DNAStringSet from Biostrings}

\item{motif}{single string, containing only DNA IUPAC characters}
}
\value{
two columns , the first being the instance the second being the
distance
}
\description{
Generate a list of locations for a single motif.
}
\examples{
## load fasta file
load(system.file("exdata","fasta.RData",package="mT1"))
pdf<-motifPDF(fasta,"CANNTG")
x<-seq(1:100)
y<-combHeights(x,pdf[,2])[[1]]
plot(x,y,main="CANNTG",ylab="Frequency",xlab="Index")
}

