% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mT1_fasta}
\alias{mT1_fasta}
\title{Fasta}
\format{A DANStringSet with 5670 members:}
\usage{
mT1_fasta
}
\description{
The nucleotide values for the genomic locations defined by peaks
}
\keyword{datasets}
