% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mT1_jaspar}
\alias{mT1_jaspar}
\title{Jaspar}
\format{A List with 4 members:
\describe{
  \item{jnames}{The composite motifs}
  \item{jfid}{The names of the moitfs, same order as jnames}
  \item{jl}{The numner of times a motif occured in the sample fasta file}
  \item{jsublM}{jnames sorted by jl}
}}
\usage{
mT1_jaspar
}
\description{
A ist containing composite motifs and names for the jaspar database
}
\keyword{datasets}
