% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mT1_peaks}
\alias{mT1_peaks}
\title{peaks}
\format{A data frame with 5670 rows and 3 variables:
\describe{
  \item{chr}{chromosome}
  \item{start}{starting base pair location}
  \item{end}{ending base pair location}
}}
\usage{
mT1_peaks
}
\description{
A set of TAL1 peaks combined from 22 Hematopoietic and Endothelial
datasets
}
\keyword{datasets}
