% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mT1_sampleMT1}
\alias{mT1_sampleMT1}
\title{objMT1}
\format{\preformatted{List of 9
 $ diff  :List of 66
  ..$ : int [1:6918, 1:2] 4 7 13 13 13 14 14 14 15 15 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:7261, 1:2] 1 1 3 3 3 3 3 3 3 3 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:6918, 1:2] 4 7 13 13 13 14 14 14 15 15 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:6207, 1:2] 1 1 1 1 1 7 7 10 10 10 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:5562, 1:2] 8 8 8 8 8 8 8 11 16 16 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:5532, 1:2] 1 1 1 1 1 1 1 1 1 1 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:5768, 1:2] 5 5 6 6 6 8 8 8 8 8 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:4612, 1:2] 3 3 3 3 3 4 10 10 10 10 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:4606, 1:2] 3 3 3 3 3 4 10 10 10 10 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:4436, 1:2] 3 3 3 3 3 5 5 6 6 6 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:3425, 1:2] 6 6 6 14 14 14 16 16 22 22 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:5651, 1:2] 4 4 15 16 18 18 18 21 25 25 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1982, 1:2] 22 22 31 31 34 34 39 39 39 39 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2021, 1:2] 4 4 4 7 7 13 13 14 21 27 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2169, 1:2] 18 21 25 26 26 37 38 38 40 50 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1729, 1:2] 4 7 15 18 21 32 40 42 48 48 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1853, 1:2] 8 22 22 27 27 38 38 40 43 55 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1415, 1:2] 13 15 18 26 39 39 39 41 41 42 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1532, 1:2] 4 27 27 34 34 37 39 39 39 42 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1324, 1:2] 8 22 22 38 38 55 57 57 57 57 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1286, 1:2] 14 16 25 37 41 47 48 60 71 71 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:5651, 1:2] 4 4 15 16 18 18 18 21 25 25 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:4692, 1:2] 1 3 3 3 4 11 11 18 18 18 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:4695, 1:2] 8 11 18 18 18 18 18 18 21 26 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:4741, 1:2] 1 1 1 1 3 3 4 4 5 5 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:4677, 1:2] 3 3 3 5 8 8 8 8 8 9 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:3954, 1:2] 4 4 10 10 10 15 17 18 18 18 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:3757, 1:2] 3 3 3 4 4 12 13 13 18 18 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:3530, 1:2] 3 3 3 5 8 8 8 8 8 12 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:3447, 1:2] 16 21 22 25 25 26 30 30 30 34 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2021, 1:2] 4 4 4 7 7 13 13 14 21 27 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2169, 1:2] 18 21 25 26 26 37 38 38 40 50 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1729, 1:2] 4 7 15 18 21 32 40 42 48 48 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1853, 1:2] 8 22 22 27 27 38 38 40 43 55 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1415, 1:2] 13 15 18 26 39 39 39 41 41 42 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1532, 1:2] 4 27 27 34 34 37 39 39 39 42 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1324, 1:2] 8 22 22 38 38 55 57 57 57 57 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1286, 1:2] 14 16 25 37 41 47 48 60 71 71 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1625, 1:2] 20 21 23 28 33 33 33 36 36 37 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1701, 1:2] 2 4 4 4 7 7 12 15 21 36 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1609, 1:2] 19 19 23 23 23 23 23 28 43 44 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1308, 1:2] 3 4 12 13 13 20 27 27 37 42 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1321, 1:2] 3 4 4 4 10 15 17 18 20 27 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1098, 1:2] 3 19 19 27 43 44 51 52 57 78 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:938, 1:2] 14 21 37 78 81 82 82 95 95 98 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1318, 1:2] 11 11 18 18 21 23 30 30 31 36 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1518, 1:2] 16 24 28 38 38 38 38 40 44 47 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1119, 1:2] 18 18 26 26 30 30 34 46 50 52 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1170, 1:2] 20 30 37 41 41 46 46 50 62 70 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1094, 1:2] 16 24 38 38 38 38 44 47 52 57 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1122, 1:2] 25 30 34 37 38 38 45 45 47 71 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1334, 1:2] 12 23 33 33 40 44 48 48 48 60 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1203, 1:2] 3 10 12 15 18 50 53 53 56 56 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1168, 1:2] 3 4 12 15 42 50 60 62 70 71 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:873, 1:2] 3 12 44 48 64 64 70 73 79 85 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:861, 1:2] 22 50 60 70 81 81 108 108 115 115 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:996, 1:2] 3 27 27 27 27 27 27 52 52 60 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1056, 1:2] 3 12 26 27 27 27 27 52 52 56 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1273, 1:2] 27 27 38 38 47 48 48 55 55 64 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1019, 1:2] 26 38 38 48 48 48 48 48 48 56 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:456, 1:2] 39 43 46 50 60 70 103 106 127 128 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:649, 1:2] 27 27 43 52 52 85 93 103 121 128 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:763, 1:2] 34 41 41 50 70 71 82 93 122 122 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:721, 1:2] 3 12 26 43 52 52 69 73 74 76 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:807, 1:2] 26 30 37 50 56 56 69 70 74 113 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:758, 1:2] 26 38 38 48 48 69 70 74 78 87 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
 $ dens  :List of 12
  ..$ : int [1:10755, 1:2] 1 1 1 1 1 2 2 2 3 3 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:3438, 1:2] 4 7 13 14 15 16 18 21 22 22 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:3895, 1:2] 3 3 3 7 10 10 10 12 13 13 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:3438, 1:2] 4 7 13 14 15 16 18 21 22 22 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2973, 1:2] 1 7 7 10 11 12 13 13 14 15 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2637, 1:2] 8 11 16 18 18 20 21 23 24 26 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2673, 1:2] 1 1 1 2 4 5 10 11 11 14 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2640, 1:2] 5 6 8 9 12 16 22 23 24 26 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2223, 1:2] 3 4 10 12 13 15 17 18 20 26 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2223, 1:2] 3 4 10 12 13 15 17 18 20 26 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:2031, 1:2] 3 5 6 8 12 16 22 24 26 27 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
  ..$ : int [1:1921, 1:2] 6 14 16 21 22 25 26 30 34 37 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "loc" "pos"
 $ sig   : logi [1:66] FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ motifs: chr [1:12] "CANNTG" "HGATAA" "MCWTNT" "HGATAA" ...
 $ combs : int [1:66, 1:2] 1 1 1 1 1 1 1 1 1 1 ...
 $ width : int 300
 $ hs    :List of 66
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 1 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 1 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 1 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 1 0 0 1 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 1 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 1 1 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 1 0 1 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 1 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 1 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 1 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
 $ mp    :List of 66
  ..$ : num [1:599] 8.87e-07 1.73e-06 1.39e-06 1.46e-06 2.44e-06 ...
  ..$ : num [1:599] 1.59e-06 2.27e-06 1.74e-06 2.10e-06 4.34e-06 ...
  ..$ : num [1:599] 8.87e-07 1.73e-06 1.39e-06 1.46e-06 2.44e-06 ...
  ..$ : num [1:599] 1.03e-06 2.00e-06 1.61e-06 1.69e-06 6.00e-06 ...
  ..$ : num [1:599] 1.16e-06 1.07e-06 7.21e-07 1.16e-06 8.15e-07 ...
  ..$ : num [1:599] 1.15e-06 3.41e-06 4.06e-06 3.72e-06 8.62e-06 ...
  ..$ : num [1:599] 0.00 0.00 1.15e-06 1.06e-06 1.13e-05 ...
  ..$ : num [1:599] 0.00 0.00 0.00 1.37e-06 1.11e-05 ...
  ..$ : num [1:599] 0.00 0.00 0.00 1.37e-06 1.11e-05 ...
  ..$ : num [1:599] 0.00 1.51e-06 1.38e-06 9.37e-07 3.05e-06 ...
  ..$ : num [1:599] 0.00 0.00 1.58e-06 3.08e-06 2.48e-06 ...
  ..$ : num [1:599] 4.70e-07 6.05e-07 6.72e-07 8.73e-07 1.88e-06 ...
  ..$ : num [1:599] 2.28e-07 4.56e-07 4.56e-07 6.08e-07 1.06e-06 ...
  ..$ : num [1:599] 2.64e-07 5.29e-07 5.29e-07 7.05e-07 2.29e-06 ...
  ..$ : num [1:599] 2.98e-07 1.99e-07 2.98e-07 3.98e-07 4.97e-07 ...
  ..$ : num [1:599] 2.95e-07 9.84e-07 1.28e-06 1.48e-06 3.44e-06 ...
  ..$ : num [1:599] 0.00 0.00 2.96e-07 1.97e-07 3.85e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 3.53e-07 3.53e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 3.53e-07 3.53e-06 ...
  ..$ : num [1:599] 0.00 3.88e-07 2.58e-07 3.88e-07 1.03e-06 ...
  ..$ : num [1:599] 0.00 0.00 4.07e-07 8.13e-07 8.13e-07 ...
  ..$ : num [1:599] 5.37e-07 1.88e-06 2.49e-06 1.95e-06 2.69e-06 ...
  ..$ : num [1:599] 6.23e-07 2.18e-06 2.88e-06 2.26e-06 5.21e-06 ...
  ..$ : num [1:599] 7.03e-07 1.67e-06 1.49e-06 9.67e-07 1.67e-06 ...
  ..$ : num [1:599] 6.95e-07 3.21e-06 5.73e-06 5.82e-06 8.43e-06 ...
  ..$ : num [1:599] 0.00 0.00 6.97e-07 1.65e-06 8.54e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 8.32e-07 8.53e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 8.32e-07 8.53e-06 ...
  ..$ : num [1:599] 0.00 9.13e-07 2.17e-06 1.94e-06 2.28e-06 ...
  ..$ : num [1:599] 0.00 0.00 9.58e-07 3.35e-06 4.43e-06 ...
  ..$ : num [1:599] 2.64e-07 5.29e-07 5.29e-07 7.05e-07 2.29e-06 ...
  ..$ : num [1:599] 2.98e-07 1.99e-07 2.98e-07 3.98e-07 4.97e-07 ...
  ..$ : num [1:599] 2.95e-07 9.84e-07 1.28e-06 1.48e-06 3.44e-06 ...
  ..$ : num [1:599] 0.00 0.00 2.96e-07 1.97e-07 3.85e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 3.53e-07 3.53e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 3.53e-07 3.53e-06 ...
  ..$ : num [1:599] 0.00 3.88e-07 2.58e-07 3.88e-07 1.03e-06 ...
  ..$ : num [1:599] 0.00 0.00 4.07e-07 8.13e-07 8.13e-07 ...
  ..$ : num [1:599] 1.61e-06 3.46e-07 8.07e-07 1.15e-06 5.76e-07 ...
  ..$ : num [1:599] 1.60e-06 3.76e-06 3.42e-06 3.42e-06 1.08e-05 ...
  ..$ : num [1:599] 0.00 0.00 1.60e-06 3.43e-07 1.62e-05 ...
  ..$ : num [1:599] 0.00 0.00 0.00 1.91e-06 1.47e-05 ...
  ..$ : num [1:599] 0.00 0.00 0.00 1.91e-06 1.47e-05 ...
  ..$ : num [1:599] 0.00 2.10e-06 4.49e-07 1.05e-06 3.74e-06 ...
  ..$ : num [1:599] 0.00 0.00 2.20e-06 2.83e-06 1.73e-06 ...
  ..$ : num [1:599] 1.67e-06 4.63e-06 5.28e-06 4.89e-06 1.15e-05 ...
  ..$ : num [1:599] 0.00 0.00 1.68e-06 1.03e-06 1.74e-05 ...
  ..$ : num [1:599] 0.00 0.00 0.00 2.00e-06 1.63e-05 ...
  ..$ : num [1:599] 0.00 0.00 0.00 2.00e-06 1.63e-05 ...
  ..$ : num [1:599] 0.00 2.20e-06 1.35e-06 1.52e-06 3.89e-06 ...
  ..$ : num [1:599] 0.00 0.00 2.30e-06 3.90e-06 3.19e-06 ...
  ..$ : num [1:599] 0.00 0.00 1.15e-06 1.28e-06 1.30e-05 ...
  ..$ : num [1:599] 0.00 0.00 0.00 1.37e-06 1.22e-05 ...
  ..$ : num [1:599] 0.00 0.00 0.00 1.37e-06 1.22e-05 ...
  ..$ : num [1:599] 0.00 1.50e-06 1.67e-06 2.01e-06 3.51e-06 ...
  ..$ : num [1:599] 0.00 0.00 1.58e-06 3.51e-06 4.03e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 1.07e-06 9.47e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 1.07e-06 9.47e-06 ...
  ..$ : num [1:599] 0.00 1.17e-06 1.01e-06 1.01e-06 2.51e-06 ...
  ..$ : num [1:599] 0.00 0.00 1.23e-06 2.46e-06 2.28e-06 ...
  ..$ : num [1:599] 0.00 0.00 0.00 5.47e-07 6.93e-06 ...
  ..$ : num [1:599] 0.0 6.0e-07 2.0e-06 1.0e-06 2.6e-06 ...
  ..$ : num [1:599] 0.00 0.00 6.29e-07 2.94e-06 3.36e-06 ...
  ..$ : num [1:599] 0.0 6.0e-07 2.0e-06 1.0e-06 2.6e-06 ...
  ..$ : num [1:599] 0.00 0.00 6.29e-07 2.94e-06 3.36e-06 ...
  ..$ : num [1:599] 0.00 0.00 1.84e-06 2.99e-06 2.07e-06 ...
 $ pvalue:List of 66
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ : num [1:599] 0 0 0 0 0 0 0 0 0 0 ...
 - attr(*, "class")= chr "mT1"
}}
\usage{
mT1_sampleMT1
}
\description{
objMT1
}
\keyword{datasets}

