% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{plot.mT1}
\alias{plot.mT1}
\title{plot mT1}
\usage{
plot.mT1(obj, motif1 = NULL, motif2 = NULL, i = NULL)
}
\arguments{
\item{obj}{mT1 object}

\item{motif1}{motif in set analyzed}

\item{motif2}{motif in set analyzed}

\item{i}{index value}
}
\description{
Used to plot the results of mT1. Note this does not accept variable
arguments to `plot.default`.
}

