% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{combHeights}
\alias{combHeights}
\title{Combination Heights}
\usage{
combHeights(x, ...)
}
\arguments{
\item{x}{A sequence of possible values represented as indicies}

\item{...}{The vectors of values to be mapped}
}
\value{
List with length of x range
}
\description{
Determines the number of occurrences in each member ... along x.
}
\examples{
x<-seq(20)
y<-runif(200,1,20)
par(mfrow=c(1,2))
plot(y)
plot(x,combHeights(x,y)[[1]])
}
