% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{diffMotif}
\alias{diffMotif}
\title{Motif Difference}
\usage{
diffMotif(fasta, motifs, mloc, cloc, i = 1, j = 2, min = 200,
  combine = "Merged")
}
\arguments{
\item{fasta}{A DNAStringSet}

\item{motifs}{The list of motif Character Atomics}

\item{mloc}{A list of indicies  of each motif in fasta}

\item{cloc}{A list of indicies of each composite motif in fasta}

\item{i}{Motif index 1}

\item{j}{Motif index index 2}

\item{min}{numerical atomic, minimum length of mloc[[i or j]]}

\item{combine}{What to do if two motifs occur in the same region
Merged|First}
}
\value{
A numeric vector with two columns , the first being the
genomic index the second being the distance.
}
\description{
Find the distances between two motifs within a subset of the genome
represented by fasta. If the distribution of motifs are already known
look into mT1:::.motifDiffFromDens.
}
