% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{eMP}
\alias{eMP}
\title{Expected Motif Probability}
\usage{
eMP(a, b, width)
}
\arguments{
\item{a}{numerical vector PWM of motif a}

\item{b}{numerical vector PWM of motif b}

\item{width}{width of fasta indicies}
}
\value{
expectation for each motif distance
}
\description{
Determine the expected probability of finding two motifs at a specific
distance based on the empirical PDFs of the individual motifs
}

