% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{findLocs}
\alias{findLocs}
\title{Find Motif Locations}
\usage{
findLocs(fasta, mloc, cloc, n1, combine = "Merged")
}
\arguments{
\item{fasta}{A DNAStringSet}

\item{mloc}{fasta The indicies that have  motifs}

\item{cloc}{fasta The indicies that have compliment motifs}

\item{n1}{The motif name, an Atomic Character of only IUPAC characters}

\item{combine}{If two motifs are found on the same index Merged|First}
}
\value{
Two columns , the first being the instance the second being the
distance
}
\description{
Finds the position of motifs on each index of fasta. The results can
be used to build the emperical PDF of motif locations.
}
