% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{getJASPAR}
\alias{getJASPAR}
\title{Get JASPAR}
\usage{
getJASPAR(name)
}
\arguments{
\item{name}{A composite motif. IUPAC DNA characters only.}
}
\value{
The name of the motif as an atomic character, if it came
from the jaspar database, otherwise NULL.
}
\description{
A function to be used in conjunction with the mT1_jaspar.
Pass `getJASPAR` a composite motif from the jaspar database
and it will return the name of the motif.
}
\examples{
# load the required Jaspar motif data
a<-mT1_jaspar$names[1]
getJASPAR(a)
}
