% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{makeTitle}
\alias{makeTitle}
\alias{makeTitle.mT1}
\title{Make Title}
\usage{
makeTitle(x, ...)

\method{makeTitle}{mT1}(x, i, ...)
}
\arguments{
\item{x}{The obj to make title from}

\item{...}{possible properties}

\item{i}{The index of interest from print(x)}
}
\value{
An Atomic Character with motif info
}
\description{
A generic function that  Generates a title to be used for plotting
an object.
}
