% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{motifDistance}
\alias{motifDistance}
\title{Motif Distance}
\usage{
motifDistance(fasta, motif1, motif2)
}
\arguments{
\item{fasta}{A DNAStringSet with a set of genomic ranges}

\item{motif1}{A motif as an atomic Character of IUPAC characters}

\item{motif2}{A motif as an atomic Character of IUPAC characters}
}
\value{
a numeric vector represented by two columns, the first being
the index the second being the position of the motif within that index.
distance
}
\description{
Find the distances between two motifs within set of genomic ranges.
The subsets are represented by DNAStringSet from the Biostrings package.
If the index of genomic ranges for each motif is already known use
\code{\link{diffMotif}}. If the distribution of motifs are already known
look into mT1:::.motifDiffFromDens.
}
\examples{
## load fasta file
fasta<-mT1_fasta
## Find distances
distances<-motifDistance(fasta,"CANNTG","HGATAA")

## Determine the range of the motif locations and plot the
## frequency
width<-fasta@ranges@width[1] # all fasta strings should be the same width
r<-seq(-width+1,width-1)
y<-combHeights(r,distances[,2])[[1]]
plot(r,y,main="CANNTG-HGATAA")
}
