% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{motifPDF}
\alias{motifPDF}
\title{Motif PDF}
\usage{
motifPDF(fasta, motif)
}
\arguments{
\item{fasta}{A DNAStringSet from Biostrings.}

\item{motif}{An atomic Character containing only DNA IUPAC characters}
}
\value{
A numeric vector with two columns , the first being the index
the second being the distance
}
\description{
Generate a distribution of locations for a single motif within a set
of genomic regions.
}
\examples{
## load fasta file
fasta<-mT1_fasta
pdf<-motifPDF(fasta,"CANNTG")
x<-seq(1:100)
y<-combHeights(x,pdf[,2])[[1]]
## plot the results
plot(x,y,main="CANNTG",ylab="Frequency",xlab="Index")
}
