% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mT1_sampleMT1}
\alias{mT1_sampleMT1}
\title{sampleMT1}
\format{mT1 object:}
\usage{
mT1_sampleMT1
}
\description{
A sample mT1 object for use in examples.
}
\details{
It was derived using mT1_fasta and the motifs
c("CANNTG","GATAA",mT1_jaspar$jsublM[1:10])
}
\keyword{datasets}
